prepare for page_alloc redesign

+++ src/kernel/types/page.rs
@@ -1,7 +1,7 @@
 
 use core::mem::MaybeUninit;
 
-pub const PAGE_SHIFT: u8 = 12;
+pub const PAGE_SHIFT: usize = 12;
 pub const PAGE_SIZE: usize = 1 << PAGE_SHIFT;
 pub const PAGE_LOW_MASK: usize = PAGE_SIZE - 1;
 pub const PAGE_HIGH_MASK: usize = !PAGE_LOW_MASK;
